+++ elasticsearch-singledocument/src/main/java/org/vertexium/elasticsearch/ElasticsearchSingleDocumentSearchIndex.java
@@ -8,6 +8,7 @@ import org.elasticsearch.action.admin.cluster.health.ClusterHealthResponse;
 import org.elasticsearch.action.admin.cluster.health.ClusterHealthStatus;
 import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
+import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsResponse;
 import org.elasticsearch.action.admin.indices.stats.IndexStats;
 import org.elasticsearch.action.bulk.BulkItemResponse;
 import org.elasticsearch.action.bulk.BulkRequest;
@@ -20,6 +21,9 @@ import org.elasticsearch.action.update.UpdateResponse;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.client.Requests;
 import org.elasticsearch.client.transport.TransportClient;
+import org.elasticsearch.cluster.metadata.MappingMetaData;
+import org.elasticsearch.common.collect.ImmutableOpenMap;
+import org.elasticsearch.common.hppc.cursors.ObjectCursor;
 import org.elasticsearch.common.settings.ImmutableSettings;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.transport.InetSocketTransportAddress;
@@ -79,7 +83,7 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
     private static final int MAX_RETRIES = 3;
     private final Client client;
     private final ElasticSearchSearchIndexConfiguration config;
-    private final Map<String, IndexInfo> indexInfos = new HashMap<>();
+    private Map<String, IndexInfo> indexInfos;
     private int indexInfosLastSize = 0; // Used to prevent creating a index name array each time
     private String[] indexNamesAsArray;
     private IndexSelectionStrategy indexSelectionStrategy;
@@ -100,7 +104,6 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
         this.allFieldEnabled = this.config.isAllFieldEnabled(false);
         this.propertyNameVisibilitiesStore = createPropertyNameVisibilitiesStore(graph, config);
         this.client = createClient(this.config);
-        loadIndexInfos(graph);
     }
 
     protected Client createClient(ElasticSearchSearchIndexConfiguration config) {
@@ -218,7 +221,19 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
         return client.admin().indices().prepareStats().execute().actionGet().getIndices();
     }
 
-    protected void loadIndexInfos(Graph graph) {
+    void clearIndexInfoCache() {
+        this.indexInfos = null;
+    }
+
+    private Map<String, IndexInfo> getIndexInfos(Graph graph) {
+        if (indexInfos == null) {
+            indexInfos = new HashMap<>();
+            loadIndexInfos(graph, indexInfos);
+        }
+        return indexInfos;
+    }
+
+    private void loadIndexInfos(Graph graph, Map<String, IndexInfo> indexInfos) {
         Map<String, IndexStats> indices = getExistingIndexNames();
         for (String indexName : indices.keySet()) {
             if (!indexSelectionStrategy.isIncluded(this, indexName)) {
@@ -238,8 +253,55 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
             addPropertyNameVisibility(graph, indexInfo, OUT_VERTEX_ID_FIELD_NAME, null);
             addPropertyNameVisibility(graph, indexInfo, IN_VERTEX_ID_FIELD_NAME, null);
             addPropertyNameVisibility(graph, indexInfo, EDGE_LABEL_FIELD_NAME, null);
+            loadExistingMappingIntoIndexInfo(graph, indexInfo, indexName);
             indexInfos.put(indexName, indexInfo);
         }
+    }
+
+    private void loadExistingMappingIntoIndexInfo(Graph graph, IndexInfo indexInfo, String indexName) {
+        try {
+            GetMappingsResponse mapping = client.admin().indices().prepareGetMappings(indexName).get();
+            for (ObjectCursor<String> mappingIndexName : mapping.getMappings().keys()) {
+                ImmutableOpenMap<String, MappingMetaData> typeMappings = mapping.getMappings().get(mappingIndexName.value);
+                for (ObjectCursor<String> typeName : typeMappings.keys()) {
+                    MappingMetaData typeMapping = typeMappings.get(typeName.value);
+                    Map<String, Map<String, String>> properties = getPropertiesFromTypeMapping(typeMapping);
+                    if (properties == null) {
+                        continue;
+                    }
+
+                    for (Map.Entry<String, Map<String, String>> propertyEntry : properties.entrySet()) {
+                        String rawPropertyName = propertyEntry.getKey();
+                        loadExistingPropertyMappingIntoIndexInfo(graph, indexInfo, rawPropertyName);
+                    }
+                }
+            }
+        } catch (IOException ex) {
+            throw new VertexiumException("Could not load type mappings", ex);
+        }
+    }
+
+    @SuppressWarnings("unchecked")
+    private Map<String, Map<String, String>> getPropertiesFromTypeMapping(MappingMetaData typeMapping) throws IOException {
+        return (Map<String, Map<String, String>>) typeMapping.getSourceAsMap().get("properties");
+    }
+
+    private void loadExistingPropertyMappingIntoIndexInfo(Graph graph, IndexInfo indexInfo, String rawPropertyName) {
+        Matcher m = ElasticsearchSingleDocumentSearchIndex.PROPERTY_NAME_PATTERN.matcher(rawPropertyName);
+        if (!m.matches()) {
+            return;
+        }
+
+        String propertyName = m.group(1);
+        String visibilityHash = m.group(2);
+        Visibility propertyVisibility;
+        if (visibilityHash == null) {
+            propertyVisibility = null;
+        } else {
+            visibilityHash = visibilityHash.substring(1); // stop leading _
+            propertyVisibility = this.propertyNameVisibilitiesStore.getVisibilityFromHash(graph, visibilityHash);
+        }
+        addPropertyNameVisibility(graph, indexInfo, propertyName, propertyVisibility);
     }
 
     private PropertyNameVisibilitiesStore createPropertyNameVisibilitiesStore(Graph graph, GraphConfiguration config) {
@@ -305,7 +367,7 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
             Recurrent.run(update, retryPolicy);
 
             if (getConfig().isAutoFlush()) {
-                flush();
+                flush(graph);
             }
         } catch (Exception e) {
             throw new VertexiumException("Could not add element", e);
@@ -366,7 +428,7 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
     private String addElementTypeVisibilityPropertyToIndex(Graph graph, Element element) throws IOException {
         String elementTypeVisibilityPropertyName = deflatePropertyName(graph, ELEMENT_TYPE_FIELD_NAME, element.getVisibility());
         String indexName = getIndexName(element);
-        IndexInfo indexInfo = ensureIndexCreatedAndInitialized(indexName);
+        IndexInfo indexInfo = ensureIndexCreatedAndInitialized(graph, indexName);
         addPropertyToIndex(graph, indexInfo, elementTypeVisibilityPropertyName, element.getVisibility(), String.class, false, false);
         return elementTypeVisibilityPropertyName;
     }
@@ -772,7 +834,7 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
                 QueryBuilders.matchAllQuery(),
                 elementTypeFilterBuilder
         );
-        SearchRequestBuilder q = getClient().prepareSearch(getIndexNamesAsArray())
+        SearchRequestBuilder q = getClient().prepareSearch(getIndexNamesAsArray(graph))
                 .setQuery(queryBuilder)
                 .setSearchType(SearchType.COUNT);
 
@@ -808,13 +870,14 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
         return results;
     }
 
-    protected IndexInfo ensureIndexCreatedAndInitialized(String indexName) {
+    protected IndexInfo ensureIndexCreatedAndInitialized(Graph graph, String indexName) {
+        Map<String, IndexInfo> indexInfos = getIndexInfos(graph);
         IndexInfo indexInfo = indexInfos.get(indexName);
         if (indexInfo != null && indexInfo.isElementTypeDefined()) {
             return indexInfo;
         }
 
-        synchronized (indexInfos) {
+        synchronized (this) {
             if (indexInfo == null) {
                 if (!client.admin().indices().prepareExists(indexName).execute().actionGet().isExists()) {
                     try {
@@ -926,7 +989,7 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
         LOGGER.debug("added %d elements", totalCount);
 
         if (getConfig().isAutoFlush()) {
-            flush();
+            flush(graph);
         }
     }
 
@@ -941,11 +1004,12 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
     }
 
     @Override
-    public void flush() {
-        client.admin().indices().prepareRefresh(getIndexNamesAsArray()).execute().actionGet();
+    public void flush(Graph graph) {
+        client.admin().indices().prepareRefresh(getIndexNamesAsArray(graph)).execute().actionGet();
     }
 
-    protected String[] getIndexNamesAsArray() {
+    protected String[] getIndexNamesAsArray(Graph graph) {
+        Map<String, IndexInfo> indexInfos = getIndexInfos(graph);
         if (indexInfos.size() == indexInfosLastSize) {
             return indexNamesAsArray;
         }
@@ -997,7 +1061,7 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
     public IndexInfo addPropertiesToIndex(Graph graph, Element element, Iterable<Property> properties) {
         try {
             String indexName = getIndexName(element);
-            IndexInfo indexInfo = ensureIndexCreatedAndInitialized(indexName);
+            IndexInfo indexInfo = ensureIndexCreatedAndInitialized(graph, indexName);
             for (Property property : properties) {
                 addPropertyToIndex(graph, indexInfo, property);
             }
@@ -1091,23 +1155,23 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
     }
 
     @Override
-    public synchronized void truncate() {
+    public synchronized void truncate(Graph graph) {
         LOGGER.warn("Truncate of Elasticsearch is not possible, dropping the indices and recreating instead.");
-        drop();
+        drop(graph);
     }
 
     @Override
-    public void drop() {
-        Set<String> indexInfosSet = indexInfos.keySet();
+    public void drop(Graph graph) {
+        Set<String> indexInfosSet = getIndexInfos(graph).keySet();
         for (String indexName : indexInfosSet) {
             try {
                 DeleteIndexRequest deleteRequest = new DeleteIndexRequest(indexName);
                 getClient().admin().indices().delete(deleteRequest).actionGet();
-                indexInfos.remove(indexName);
+                getIndexInfos(graph).remove(indexName);
             } catch (Exception ex) {
                 throw new VertexiumException("Could not delete index " + indexName, ex);
             }
-            ensureIndexCreatedAndInitialized(indexName);
+            ensureIndexCreatedAndInitialized(graph, indexName);
         }
     }
 
@@ -1214,8 +1278,9 @@ public class ElasticsearchSingleDocumentSearchIndex implements SearchIndex, Sear
         return config;
     }
 
-    public boolean isPropertyInIndex(String propertyName) {
-        for (Map.Entry<String, IndexInfo> entry : this.indexInfos.entrySet()) {
+    public boolean isPropertyInIndex(Graph graph, String propertyName) {
+        Map<String, IndexInfo> indexInfos = getIndexInfos(graph);
+        for (Map.Entry<String, IndexInfo> entry : indexInfos.entrySet()) {
             if (entry.getValue().isPropertyDefined(propertyName)) {
                 return true;
             }