+++ elasticsearch-singledocument/src/main/java/org/vertexium/elasticsearch/MetadataTablePropertyNameVisibilitiesStore.java
@@ -5,18 +5,22 @@ import org.vertexium.Authorizations;
 import org.vertexium.Graph;
 import org.vertexium.GraphMetadataEntry;
 import org.vertexium.Visibility;
+import org.vertexium.util.VertexiumLogger;
+import org.vertexium.util.VertexiumLoggerFactory;
 
 import java.nio.charset.Charset;
 import java.util.*;
 
 public class MetadataTablePropertyNameVisibilitiesStore extends PropertyNameVisibilitiesStore {
-    public static final String METADATA_PREFIX = "propertyNameVisibility.";
+    private static final VertexiumLogger LOGGER = VertexiumLoggerFactory.getLogger(MetadataTablePropertyNameVisibilitiesStore.class);
+    public static final String PROPERTY_NAME_VISIBILITY_TO_HASH_PREFIX = "propertyNameVisibility.";
+    public static final String HASH_TO_VISIBILITY = "visibilityHash.";
     private static final Charset UTF8 = Charset.forName("utf8");
     private Map<String, Visibility> visibilityCache = new HashMap<>();
 
     public Collection<String> getHashes(Graph graph, String propertyName, Authorizations authorizations) {
         List<String> results = new ArrayList<>();
-        String prefix = getMetadataPrefixWithPropertyName(propertyName);
+        String prefix = getPropertyNameVisibilityToHashPrefix(propertyName);
         for (GraphMetadataEntry metadata : graph.getMetadataWithPrefix(prefix)) {
             String visibilityString = metadata.getKey().substring(prefix.length());
             Visibility visibility = getVisibility(visibilityString);
@@ -39,22 +43,46 @@ public class MetadataTablePropertyNameVisibilitiesStore extends PropertyNameVisi
 
     public String getHash(Graph graph, String propertyName, Visibility visibility) {
         String visibilityString = visibility.getVisibilityString();
-        String metadataKey = getMetadataKey(propertyName, visibilityString);
-        String hash = (String) graph.getMetadata(metadataKey);
+        String propertyNameVisibilityToHashKey = getMetadataKey(propertyName, visibilityString);
+        String hash = (String) graph.getMetadata(propertyNameVisibilityToHashKey);
         if (hash != null) {
+            saveHashToVisibility(graph, hash, visibilityString);
             return hash;
         }
 
         hash = Hashing.murmur3_128().hashString(visibilityString, UTF8).toString();
-        graph.setMetadata(metadataKey, hash);
+        graph.setMetadata(propertyNameVisibilityToHashKey, hash);
+        saveHashToVisibility(graph, hash, visibilityString);
         return hash;
     }
 
-    private String getMetadataPrefixWithPropertyName(String propertyName) {
-        return METADATA_PREFIX + propertyName + ".";
+    private void saveHashToVisibility(Graph graph, String hash, String visibilityString) {
+        String hashToVisibilityKey = getHashToVisibilityKey(hash);
+        String foundVisibilityString = (String) graph.getMetadata(hashToVisibilityKey);
+        if (foundVisibilityString == null) {
+            graph.setMetadata(hashToVisibilityKey, visibilityString);
+        }
+    }
+
+    @Override
+    public Visibility getVisibilityFromHash(Graph graph, String visibilityHash) {
+        String visibilityString = (String) graph.getMetadata(getHashToVisibilityKey(visibilityHash));
+        if (visibilityString == null) {
+            LOGGER.warn("Could not find visibility matching the hash \"%s\" in the metadata table.", visibilityHash);
+            return null;
+        }
+        return new Visibility(visibilityString);
+    }
+
+    private String getHashToVisibilityKey(String visibilityHash) {
+        return HASH_TO_VISIBILITY + visibilityHash;
+    }
+
+    private String getPropertyNameVisibilityToHashPrefix(String propertyName) {
+        return PROPERTY_NAME_VISIBILITY_TO_HASH_PREFIX + propertyName + ".";
     }
 
     private String getMetadataKey(String propertyName, String visibilityString) {
-        return getMetadataPrefixWithPropertyName(propertyName) + visibilityString;
+        return getPropertyNameVisibilityToHashPrefix(propertyName) + visibilityString;
     }
 }